admin: silence compiler warnings

+++ flash-graph/vertex_index_reader.cpp
@@ -698,7 +698,7 @@ bool sparse_self_vertex_compute::run(vertex_id_t start_vid, index_iterator &it)
 {
 	assert(start_vid == get_first_vertex());
 	assert(get_last_vertex() - get_first_vertex() + 1
-			== it.get_num_vertices());
+			== (size_t)it.get_num_vertices());
 	// If #ragnes is 1, we'll do it in the dense_self_vertex_compute.
 	assert(num_ranges > 1);
